+++ eclair-demo/src/test/scala/fr/acinq/eclair/ClaimReceivedHtlcSpec.scala
@@ -60,7 +60,7 @@ class ClaimReceivedHtlcSpec extends FunSuite {
       txOut = TxOut(10 satoshi, OP_DUP :: OP_HASH160 :: OP_PUSHDATA(Crypto.hash160(Alice.finalPubKey)) :: OP_EQUALVERIFY :: OP_CHECKSIG :: Nil) :: Nil,
       lockTime = abstimeout + 1)
 
-    val sig = Transaction.signInput(tx2, 0, Script.write(htlcScript), SIGHASH_ALL, tx.txOut(0).amount.toLong, 1, Alice.finalKey)
+    val sig = Transaction.signInput(tx2, 0, Script.write(htlcScript), SIGHASH_ALL, tx.txOut(0).amount, 1, Alice.finalKey)
     val witness = ScriptWitness(sig :: Bob.R :: redeemScript :: Nil)
     val tx3 = tx2.copy(witness = Seq(witness))
 
@@ -71,10 +71,10 @@ class ClaimReceivedHtlcSpec extends FunSuite {
     val tx2 = Transaction(
       version = 2,
       txIn = TxIn(OutPoint(tx, 0), Array.emptyByteArray, reltimeout + 1) :: Nil,
-      txOut = TxOut(10 satoshi , OP_DUP :: OP_HASH160 :: OP_PUSHDATA(Crypto.hash160(Bob.finalPubKey)) :: OP_EQUALVERIFY :: OP_CHECKSIG :: Nil) :: Nil,
+      txOut = TxOut(10 satoshi, OP_DUP :: OP_HASH160 :: OP_PUSHDATA(Crypto.hash160(Bob.finalPubKey)) :: OP_EQUALVERIFY :: OP_CHECKSIG :: Nil) :: Nil,
       lockTime = abstimeout + 1)
 
-    val sig = Transaction.signInput(tx2, 0, Script.write(htlcScript), SIGHASH_ALL, tx.txOut(0).amount.toLong, 1, Bob.finalKey)
+    val sig = Transaction.signInput(tx2, 0, Script.write(htlcScript), SIGHASH_ALL, tx.txOut(0).amount, 1, Bob.finalKey)
     val witness = ScriptWitness(sig :: Hash.Zeroes :: redeemScript :: Nil)
     val tx3 = tx2.copy(witness = Seq(witness))
 
@@ -82,7 +82,7 @@ class ClaimReceivedHtlcSpec extends FunSuite {
   }
 
   test("Blob can spend this HTLC right away if he knows the revocation hash") {
-    val sig = Transaction.signInput(tx1, 0, Script.write(htlcScript), SIGHASH_ALL, tx.txOut(0).amount.toLong, 1, Bob.finalKey)
+    val sig = Transaction.signInput(tx1, 0, Script.write(htlcScript), SIGHASH_ALL, tx.txOut(0).amount, 1, Bob.finalKey)
     val witness = ScriptWitness(sig :: Bob.revokeCommit :: redeemScript :: Nil)
     val tx2 = tx1.copy(witness = Seq(witness))
     Transaction.correctlySpends(tx2, Seq(tx), ScriptFlags.STANDARD_SCRIPT_VERIFY_FLAGS)
